Let users reach app details from ResolverActivity
A long press on a list item in ResolverActivity (shown when users are
asked to choose an activity to complete an action) will now open
the app detail settings for the package providing that item.

This is a work in progress; awaiting more UX input on other changes
and better discoverability of this feature.

Change-Id: I2cc07c5520fdd23cf1a43f169b114295f5d9d0ac

Former-commit-id: 2d80962b2e93ae4e59d0197487bd0f6aad4cf26e