Do not enable WAL mode for the WebView database.
There is no need to enable WAL for the WebView database because it
does not require high throughput.

Enabling WAL increases memory usage and can also cause problems if
the same database is opened in multiple processes, which can happen
if an application has UI code running in multiple processes.  Normally,
we would use a content provider to ensure that the database is only
opened within one process (and then accessed remotely view RPC), but
we can't do that for WebView because it is part of the framework.

Deleted the call to setLockingEnabled.  It does nothing.

Bug: 6124556
Change-Id: Ic397c74872fe6b1246da0c859eab0d5147045c86

Former-commit-id: 9ac2d3bcde40abce8ab8af56560ea6b3ee5f9941