Fix NDEF documentation to indicate the message may be null.
The current NFC stack formats tags to the INITIALIZED state
as defined by NFC forum; in that state the tag has the
NDEF Capability Container, but does not contain any message
yet.

Tags in that state (correctly) return the NDEF technology,
but the documentation does not specify that the message
may be null.

Also, get rid of buggy getLastErrorCode and use
(cached) presence check value to determine if tag was
lost during read.

Change-Id: If4293428093024ba9cda5dd7c9979b8b06353234

Former-commit-id: a032783241cbbed47ed05df32c56298ee0f9902b