Fix key handling
putListener() returns 0 when the argument is null. And The key value of listener
registered first is always 0. For this reason, if the p2p functions are called continuously
and first call is without listener and the second call is with listener, then the message
against first call wrongly pick up the second call's listener because the key value is
the same. In order to avoid this issue, we don't use 0 as the valid listener key.

Change-Id: I0cc960b2ad37f17cf7f528d839b39aa272b83670
Signed-off-by: Yoshihiko Ikenaga <yoshihiko.ikenaga@jp.sony.com>

Former-commit-id: b5e00bda40398b8738192abfe9b745464a54865a