Fix 5849324: Changed FUL extents for pattern backup
The extents of FUL when using PIN unlock are such that there are
4 dps of space between FUL and the adjacent text above and below.
The extents of FUL when using pattern backup were such that FUL
was very far below the adjacent text above.  Pattern backup is
now consistent with PIN backup - 4 dps of space between FUL and
the adjacent text above and below.

Note that it is not possible for FUL to be the same exact size
when using pattern backup vs. PIN backup because the rest of
the layout is different.  Making it a consistent distance from
the surrounding text is the best we can do.

To fix this, the FUL area simply had to cover an extra row in the
grid layout (a space view), and the top and bottom margins had to
be set to 4 dp.  Since only the FUL view is affected, it does not
impact the rest of the layout.

This has been tested on both a Prime and Crespo to make sure it
works across different size devices.  All lock methods have been
tested to make sure other methods were not affected.

Change-Id: I518f1b616cf7bedc510979d422ebc72e07a5aff4

Former-commit-id: cf8d9073a66072a5a94ebbe24003eb63c18906f5