Fix AnimatorSet duration issue
Setting the duration on an AnimatorSet should propagate that value
to its children. This works, but only if all children are added to the set
before setDuration() is called. This fix delays that propagation until
the set is started, making it possible to have a more flexible order of when
the children are added and when the duration is set.

Issue #6324904 AnimatorSet durations too long

Change-Id: I797971c2310eb2e3fe931b4aa35de505f2a519f7

Former-commit-id: c299a3384171e36fc9ab6d1011d8a589a7f344d1