Ensure that touch and hover targets are cleared when needed.
When views are removed from a view or a view is detached from
a window, we need to update the touch and hover targets appropriately.

Failing to do this resulted in a NPE while dispatching an
ACTION_HOVER_EXIT to a view that had previously been removed.
Removed views should not get input events.

Change-Id: I4af4f8e2c4028347d3f570894fd1b3b366d11455

Former-commit-id: 59a422e90035ce5df45c526607db2d3303e3112e