Pause the WebCore thread watchdog at the right time.
There are times when we deliberately put the WebCore thread
to sleep (i.e. waiting for a user to complete a JS prompt or
dismiss an alert). In this case, we don't want the WebCore thread
watchdog to fire. Pause the watchdog before we put the WebCore thread
to sleep and resume it again when the thread wakes up.

Factor out the repeated send/pause/wait/resume code into it's own
function.

Bug: 6366520
Change-Id: I17ecf9d466ce21b25a9e5cb8ff4cb0e5fab7605b

Former-commit-id: cba778dc21ed13711ec020c12a64836bce3c2742