Provide (general) mecahnism for users of GridLayout to either suppress or react to diagnostics.
GridLayout uses a constraint solver to perform its layout operation. Some layout problems
result in systems of constraints that are inconsistent and so cannot be solved. GridLayout
detects this condition and makes corrections so as to always guarantee that the layout
operation will complete. In those cases where corrections are made, GridLayout issues a
warning to the Log. This CL allows (internal) clients of GridLayout to both swtich off
the logging and/or detect when this condition is encountered.

Change-Id: I5b871003381f81cf0a76ad3de767e7f8b8349923

Former-commit-id: 211d03322cf6e8a6c8a947bf8a494abd78a767c2