Add Droid Devanagari/Tamil fonts.
These fonts are replacements for the Lohit Devanagari/Tamil fonts.
We need to fit Devanagari onto all builds, so we only provide a single,
regular-weight font there, but Tamil is omitted for SMALLER_FONT_FOOTPRINT
builds and gets UI/non-UI and Regular/Bold weight versions. The UI versions
of the fonts are used for UI/system display, and the non-UI versions are
used only by WebView.

Bug: 6318791
Change-Id: I50ff6ec4bb428c0ac30049273f03a94de05b0c4f

Former-commit-id: bd1844d272afd7416a37765b90ecd843c0b99c9f