Fix 6395288: Added lock to avoid unbind race condition
If turning the power off while FUL was closing (due to a timeout or a
cancel for example), it would be possible for unbind() to get called
twice due to a race condition.  Turning the phone off calls unbind()
from the UI thread, while the other close comes from a binder thread
since it is coming from the service.

PATCH SET 1 attempted to solve the problem by adding a lock, but
having a lock around the bind is a bit scary.

PATCH SET 2 takes a new approach by having all calls coming from a
binder thread to be sent as messages to be handled on the UI thread.
Having all events occur on the UI thread removes the possibility of
race conditions and makes the code stronger by making everything
happen in a deterministic order.

This commit also cleans up the logcat logging a bit.  A couple of the
log messages are now printing without DEBUG being set to true.  This
is by no means spamming the log and they of course are only logged
when FUL is being used.  But it serves to give us some meaningful
information from bug reports that are currently showing nothing.  The
statements that are now logged would have made this particular bug
easy to track down.

Change-Id: I25a65c0808d88cb941439e5bf1f989dba8608be4

Former-commit-id: 257f2ecc97d294e95b069547466d2054926d960f