Fix data roaming notification.
We shouldn't tell people we're not connected due to roaming if it's really because
of data-disabled.

Use the data-disabled setting to decide if we should inform people about
data-discon due to roaming.  Note this doesn't effect other notifications about roaming.

Also fix the data-enable/disable code to take roaming into account and clear/set the
roaming notification as needed (if data is enabled while we can't get on due to roaming
we should put up the roaming notifcation..  if data is disabled while that notification is
up we should take it down).

bug:5805666
Change-Id: I5c925dfdca4abc27e0034b113508df5719f04fae

Former-commit-id: 4dafaf117e680c138b3b0e0893a8ce09bc75f519