Fix flashy transition in QuickContacts
Calls to relayout were forcing outgoing app reported visibility to
false. Because there was a DummyAnimation in the outgoing app the
stepAnimation was forcing the app Transformation alpha value to 0
based on the most recent reportedVisibile value. This was causing the
outgoing app to disappear for an instant.  Moving the visibility test
to the time at which the DummyAnimation is set fixes this problem.

Fixes bug 5908102.

Change-Id: Ib574728a007a0af759990816db42e23ba315b468

Former-commit-id: 94ef9df22c5960b74359e1a98e5c210787d382f2