Fix 6437474: Fixed black box appearing on rotation
The biometric unlock initializeView() function is called every time
the lockscreen is recreated.  Since this normally happens when the
device turns off, initializeView() was covering the backup lock so the
backup lock is not exposed when the device turns back on.  However,
initializeView() is also called when lockscreen is recreated due to an
orientation change.

With this change, the show() call to cover the backup lock has been
moved out of initializeView(), and the backup lock is now only covered
when the screen is turning off, preventing the backup lock from being
covered on an orientation change.

This also includes changes to prevent biometric unlock function calls
from occurring when SIM or Account unlock is in use.  In fact, in any
situation where we know FUL won't be used, we don't even construct it.
This is not only more efficient, but it also cuts down on the
possibility of errors where FUL is being used when it shouldn't be.

Change-Id: Ie97761840df8de5701703d9b9b991726fb601064

Former-commit-id: 88240596fce58afe79a4d06752cd8f6c086e2600