Finish up issue #6249094: Display permissions based on relevance to private data
Added bitmaps for permission group icons.  We had to pick either
icons for a dark or light background, and it had to be light for
settings, so the installer theme has gone back to dark.

Tweaked the permission groups to match the icons.

Redid the group priorities to make them correct (they were written
with higher priorities being less important).

Fixed bug where priority was not being used for system apps.

Change-Id: Ia3d3f0c1614a11b4cfd11682ce9e6c9f87d75c06

Former-commit-id: 99222d212f9ff5081d4ce6eef09dbe8eff85b83a