Fix 6398209: Improve responsiveness of swipe up to search
This fixes an issue where the swipe-up gesture for search was broken
on phones and tablets. Because the underlying window was slippery,
there was a race condition between the touch moving outside the window
and the layout flag change in the search panel being noticed.

As a result, the code would sometimes inadvertently dismiss
the search panel even though the gesture was successful.

It also changes the timing slightly so we show the search
panel longer for quick gestures.

Change-Id: I30c04b21d3367db4d41c01f23e27edf366711462

Former-commit-id: d99e7fd9fadfa46b6732c21e6a7a54755ceeb3c3