Minor changes in SpellCheckSpan pool management in SpellChecker
Bug 6464190

The 'inProgress' flag is set to false when the SpellCheckSpan starts
to get used (instead of a less intuitive when it is removed).

Pool recycling in handled by onSpellCheckSpanRemoved, called from
the TextView's SpanWatcher, when a SpellCheckSpan is removed for any
reason (from the SC code or due to text editing).

The other change is that Sentence SC now correctly removes the span
from the text (and hence recycles it in the pool).

Change-Id: If8b433fd5e41d4dc0304a127ebcc088ea1eecaa7

Former-commit-id: 69865bd6860a97793a06523a48dfe6472e9b7562