changed pokeWakelock() callback to take time in milliseconds
Changed the pokeWakelock() call back to take one argument - the duration to stay awake in ms.  This
change was needed in order to poke the wakelock for the duration of the watchdog timeout.  This
must be done in the service because the duration of the watchdog timeout is unknown at this point.

Moved pokeWakelock() from start() to handleServiceConnected() to make sure that this poke happens
before the poke in the service.  This poke is still needed to account for when devices are rotated.

Change-Id: I19d62df1489514de0588ebb937678358e70ffc95

Former-commit-id: dff307697d7a7482efe4c10cb3b07b9249524a42