Fix 6040870: Update MultiWaveView for new device
This change allows more flexibility in target placement in MultiWaveView.
It now supports a new way of specifying chevron drawables that allows
them to be moved in directions corresponding to more than the four basic
directions (right, up, left, down).

Chevron drawables can now be updated in an overlay.

This change also adds a place holder and makes a minor tweak to the layout
on 720dp devices where the navbar buttons weren't centered.

Change-Id: Icd319ec5f276870380e27737c873e78f599ff751

Former-commit-id: 4c351d62e7a09bcc29e7d0329bcdd947a302cf40