Fix 6485058: remove OnTriggerListeners from Wave*Methods in cleanup
This fixes a rare crash that could happen when the device orientation
changes while the handle is held over a target.  LockScreen.cleanUp()
was being called which set mCallback to null which then causes it
to crash in MultiWaveViewMethods.onTrigger().

The solution is to also remove OnTriggerListeners in LockScreen.cleanup().

Change-Id: I487c7c9dbbe40417e685b39f2e79b1c43b30fa00

Former-commit-id: 8818edbe2f5bef9cf161ef57e9296ff302ad44f6