Fix 6483412: Disable search from navbar while SUW is running
This fixes a problem where navbar gestures were enabled during
SetupWizard.

There's no clear signal for handling this in StatusBar since there
are a number of flags that we might use to determine the availability
of the search gesture.

However, the UI design generally calls for the gesture to be available
when the home button is visible, so I think it makes sense to start
with that and enable others as we encounter issues.

Change-Id: I7b37964ae5b51bf803d5808605f1afe694a88464

Former-commit-id: b4238e0c59d8e1890f7752934bdf4d52fa89bf1b