Modify auto brightness to return to dim level.
Previous auto brightness policy was to only ratchet screen brightness
up and never return to dim levels until the device was turned off. This
caused the screen to be very bright in dim areas and reduced battery
life. This change allows the screen to dim down over the course of
a minute to the desired brightness level.

Fixes bug 6422464.

Change-Id: I5592c06a495650c6496ec2fc9846d7a10fac4b16

Former-commit-id: 196943fb51789e7c3a39ae7d56aceba5e1304137