Streamline the logic to add account.
This cl ensures that we immediately route the user to the add account
activity if they don't have an account and their is only one relevant
account type. Also reordered the setContent logic to reduce flicker.

Note that as of this CL there is still some flicker remaining when
launching G+ without an account. But it appears to be fixed in other
apps.

Bug: 6455975
Change-Id: I91e33b4fb9618a31765b4a8651334b1c52640828

Former-commit-id: cf0a881f1c27718f686a307e6c94213815ee9dc1