Fix issue #5680541: onStartInputView called upon focus loss
We should tell the app that it is inactive, before unbinding.
Otherwise when it is told to unbind it will see that it is still
supposed to be active and immediately re-bind.

Also change the calls to set the active state to go through the
message dispatch path, to ensure ordering is correct.

Change-Id: I246241eac8f7521f42c4c1eee7f46097337e7303

Former-commit-id: a6e41342e2159402e33866e7145be357065d9c9a