Fix 6449259: FUL no longer overlaps music controls
When Face Unlock was used with pattern backup on a Prime, it was
partially covering the lockscreen music controls that appear when
music is playing.

This makes Face Unlock take up slightly less area when using pattern
backup on phones in portrait layout, regardless of whether the music
controls are showing.  Eventually I would like to make it still use
the full area when the music controls aren't showing, but this is the
safest possible fix for JB and looks ok.

Change-Id: Ib2af440ed0f2729dfff9b58cd154efad43ef5499

Former-commit-id: bad20898f5309bbc488ece65edd5f50443c37256