Fallback to fallbackring if ringtone can't be played.
If the specified ringtone file exists but cannot be played
the Ringtone will try to play the ringtone in a remote player.
But since the file is corrupt it will not be played in the
remote player either.
According to the documentation it should attempt to
fallback on another sound. This commit will do that.

Change-Id: I0216d61ca874eef0f168ad4d5bfb07491e01e654

Former-commit-id: 1c25a21ffcfbc78a0edced4076d533f34902428a