Get a WebViewDatabaseClassic instance lazily. Do not merge.
Rather than caching the WebViewDatabaseClassic instance in a member,
grab the instance when we need it.

In the case of WebSyncManager - neither it or it's subclass uses
the WebViewDatabaseClassic (and it's subclass is final) so remove
it's initialisation. We must leave the member declared however
to satisfy the published API.

Bug: 6482144
Change-Id: Ibf5186a2851d17f5ad62b35ad543aeadff5618d9

Former-commit-id: 078b5913799610c3853675508e7ccc02c45f56ba