Skip LayoutTransition animations on objects of size (0,0)
LayoutTransition runs changing animations on all objects that change between
now and the next layout. This works in most normal situations, but when a container
is becoming visible, or being added to its container, or other first-time situations,
then some of the views and parent hierarchy may be of size (0,0). The user really
shouldn't need to see an animation up from these nonsense values, so we just
skip running the animation on these objects and simply place the objects where they
need to go.

Issue #6597648 view should not animate up from size (0,0)

Change-Id: I2c355a68bf1ce3b41fbec01ad95c78d83562ba32

Former-commit-id: 66ef1a201ea9df71a8ec9b2d1aaab1eb1180ae40