Bug 6605167: Turning off FUL during error counts as attempt
If FUL was turned off while the error message was displayed it would
not count as a failed attempt.  This commit changes the
reportFailedAttempt callback to just report the failure rather than
report the failure and close.  This allows the FUL service to send the
reportFailedAttempt message earlier and then later close using the
existing cancel function, which closes without reporting a failed
attempt.

Change-Id: Ib3b76f477a98b149fcccc32ac39ecaeccd88a7e2

Former-commit-id: 7019325e183ba1abf1869ccffc4b1ecd8c771a44