Ensure we compute display name when needed
Since the media scanner no longer uses an in memory cache, it always
found an entry for files inserted via MTP, which made it follow a
different code path than before, which in turn made the media
provider skip computing the display name column.
By clearing the row id in the media scanner when handling an mtp
request, we steer it back to the correct code path.
b/6297105
b/6525450

Change-Id: I624522b176a0057d246ce3040598af255b00a141

Former-commit-id: 81e85bbf671f228600fba059e30cb22565bcfe9d