Revert "Remove ViewTreeObserver allocations"
This reverts commit a13978137a32ac870203fdec88be9cd9a932b820 [formerly b999cc118fe430699e9a67d5dab355125b873abb].

There was an assumption in this earlier change that observer dispatching could not be
recursive - we could only ever have one iteration on the observer listener list. This
assumption broke down in a specific app, and maybe in more, so reverting the change for now.
We should probably find a way to accomplish the same allocation-minimizing goal without
causing exceptions when violating our assumptions.

Issue #6620795 [Application compatibility] Lufthansa app crashes

Change-Id: I1c1f9ad329c14398feb0e74ce77e1a07111f7d1f

Former-commit-id: 0f8ffd83745f718a476564f35a2f7fd4637275bc