Remove an unnecessary animation.
When the notification panel is full-width (small devices
like phones) we can just leave the icons showing all the
time, which saves a little CPU during the
performance-critical shade expand operation.

On sw600dp devices we continue to use DISABLE_SYSTEM_INFO to
smoothly hide the status bar when the notification panel
appears (and show it when the panel is dismissed).

Bug: 6555172
Change-Id: I59a1226549dd1e51c375c3972b47079c1ee856b0

Former-commit-id: 66d92f81093e648186b31d0416555ec05aff0e0b