BUG 5457035: lowering max FUL failed attempts to 3
After an unrecognized face occurs 3 times in a row, we disable FUL until the user unlocks via the
backup lock.  Lowering this values makes spoofing with liveliness enabled more difficult.  Since
we currently don't differentiate between the max number attempts with and without liveliness
enabled, we had to lower it for all uses of FUL.

Change-Id: I7a429f64cde2767ddd2ceb0885343acd0b802aac

Former-commit-id: aa24906de299b392a3aa8576d3f0df77089f53c2