Add new callback for MediaRouter to found out about device changes.
The AudioService now has an API to call to get the currently
connected devices, and later reports of changes in connection
state.  The information includes the name of the bluetooth
device if one is connected for display to the user, and states
for all of the pluggable devices.  No longer requires a Bluetooth
permission to keep the routes updated.

Change-Id: I81ca421c60592fbc1592477d59bf1c9d1b64954a

Former-commit-id: 632ca417f0a33e3fa9ccece531afa2db3f0d4a30