Fix issue #6675499: java.lang.RuntimeException: Unable to start...
...activity ComponentInfo{com.google.android.gm/
com.google.android.gm.ui.MailActivityGmail}: java.lang.NullPointerException

There were a number of places in FragmentManagerImpl where we were
not dealing with mAdded being null.  In the original implementation,
mAdded would almost always be null if mActive is null.  As we have
added features, this has become a less strong guarantee (and it actually
was never completely guaranteed), but there are a lot of places where
we would check for mActive being non-null and assume this meant
mAdded is non-null.

Fix these to correctly check for mAdded.

Bug: 6675499
Change-Id: I2a6a801d8bc89550fc73e12c9c3f8bb0ad6c7fa4

Former-commit-id: acdfbcca6f91bdf3f77ed7b2e699c97ce98c1fb8