Fix 6656710: center dot pattern in GlowPadView
This fixes a bug where the dots weren't aligned with the center of the view.
The matrix should not contain the offset of the image itself since we really
want to scale about the abstract points.  The code now subtracts the image
centering positions (cx,cy) from the point to be scaled.

Change-Id: I1d46a5d51b1fd1d39944c64baf86c306c2d5e923

Former-commit-id: 94754ef36d1ceacb9a335202585e60c82b1f4b0d